Work in progress : Three.js scene + cameras
Thinking about the range of x y z cartesian space that I want to actually
be visible in the scene. It seems like we want the intervals (0, 100) visible
on each axis, with (0, 0, 0) visible in the lower left corner. That way, we
can determine the min and max for our axes and have a really simple calculation
as far as where we draw each sphere.